refactor: extract api.js and units.js from index.js
- api.js: requestPlotData() and downloadOutputFiles() — all server fetch logic
- units.js: getBasicParameters(), refreshTimeUnitFactor(),
  applyTimeUnitToFigures(), TIME_UNIT_LABEL — time unit conversion
- Simplify addDownloadFunction to delegate to downloadOutputFiles
- Rename time unit keys: R0Cs → soundSpeed, R0Va → alfvenSpeed
  for clearer internal naming; display labels unchanged
- Remove ~140 lines of extracted code from index.js
