+++ views/plot.pug
@@ -43,8 +43,8 @@ html(lang="en")
                 div(id="unit-chooser" class="dropdown")
                     label(for="time-unit-select") Time unit
                     select(id="time-unit-select" name="time-unit")
-                        option(value="R0Cs" selected) R_0/Cs
-                        option(value="R0Va") R0/Va
+                        option(value="soundSpeed" selected) R_0/Cs
+                        option(value="alfvenSpeed") R0/Va
                         option(value="tstep") tstep
                         option(value="microsecond") μs
